% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{calc_average_wind_direction}
\alias{calc_average_wind_direction}
\title{Calculate the average wind direction.}
\usage{
calc_average_wind_direction(wdir)
}
\arguments{
\item{wdir}{in degrees (0 - 360) where 0 or 360 is from the north.}
}
\value{

}
\description{
Calculate the average wind direction.
}
