% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{calc_wind_direction_difference}
\alias{calc_wind_direction_difference}
\title{Calculate absolute difference of the two wind directions in degrees}
\usage{
calc_wind_direction_difference(x_deg, y_deg)
}
\arguments{
\item{x_deg}{wind direction in degrees}

\item{y_deg}{wind direction in degrees}
}
\value{
absolute difference of the two wind directions in degrees
}
\description{
Calculate absolute difference of the two wind directions in degrees
}
