% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{find_xy_ncdf4}
\alias{find_xy_ncdf4}
\title{Find the xy index of WRF grid}
\usage{
find_xy_ncdf4(in.ncfile, sitelon, sitelat, VAR = "no.wind", CELL_SIZE = 0.013)
}
\arguments{
\item{in.ncfile}{input NetCDF file}

\item{sitelon}{longitude for (point) site for which WRF variable is extracted}

\item{sitelat}{longitude for (point) site for which WRF variable is extracted}

\item{VAR}{variable of interest from WRF output}

\item{CELL_SIZE}{the horizontal spatial resolution of WRF grid cell.}
}
\value{

}
\description{
Find the xy index of WRF grid
}
\examples{

}
