% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_wind.R
\name{merge_wrf_with_vertical_wind_obs}
\alias{merge_wrf_with_vertical_wind_obs}
\title{Merge WRF winds with vertical wind observations matching the height (relative to the ground)}
\usage{
merge_wrf_with_vertical_wind_obs(df_wrf, df_obs, tz = "gmt")
}
\arguments{
\item{df_wrf}{data frame for WRF output}

\item{df_obs}{data frame for observation}

\item{tz}{time zone, 'gmt' is the default.}
}
\value{

}
\description{
Merge WRF winds with vertical wind observations matching the height (relative to the ground)
}
\examples{
Example of df_wrf (m_agl and gmt is required to match WRF to obs)
     gmt level m_agl      u     v  wspd  wdir
     <dbl> <chr> <dbl>  <dbl> <dbl> <dbl> <dbl>
     1 2020020100 L1     12.7 0.427  -1.91  1.96  347.
     2 2020020100 L2     39.1 0.433  -2.10  2.15  348.
}
