% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{pivot_long_wind_profiler_mat}
\alias{pivot_long_wind_profiler_mat}
\title{Combine and reformat wind profiler (915 MHz) data in matrix to long format}
\usage{
pivot_long_wind_profiler_mat(gmt, height.agl.km, wind.speed, wind.direction)
}
\arguments{
\item{height.agl.km}{height in km a.g.l. in matrix of szie (hour, height_level)}

\item{wind.speed}{wind speed (m/s) in matrix of size (hour, height_level)}

\item{wind.direction}{wind direction (meteorological direction) in matrix of size (hour, height_level)}
}
\value{

}
\description{
Combine and reformat wind profiler (915 MHz) data in matrix to long format
}
\note{
The raw data were downloaded from ftp://ftp1.psl.noaa.gov/psd2/data/realtime/Radar915/CnsWind/ and processed in a matrix.
}
