% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_wind.R
\name{pivot_long_wrf_winds}
\alias{pivot_long_wrf_winds}
\title{Combine and reformat WRF wind data in matrix to long format}
\usage{
pivot_long_wrf_winds(dt, u, v, wrf_lv_m, num_lv = 50, tz = "gmt")
}
\arguments{
\item{dt}{time (YYYYMMDDHH format), typically in GMT}

\item{u}{U component of WRF wind, m/s, with size of (length of dt, length of (num_lv)), e.g., dim (u) 434  50}

\item{v}{V component of WRF wind, m/s, with size of (length of dt, length of (num_lv)), e.g., dim (u) 434  50}

\item{wrf_lv_m}{WRF vertical winds in meters, with size of (length of dt, length of (num_lv)), e.g., dim (u) 434  50}

\item{num_lv}{number of levels, 50 levels by default}

\item{tz}{time zone, gmt by default}
}
\value{

}
\description{
Combine and reformat WRF wind data in matrix to long format
}
