% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_height_vs_wspd}
\alias{plot_height_vs_wspd}
\title{Plot height (across different vertical levels) versus wind speed with error bars.}
\usage{
plot_height_vs_wspd(
  df,
  x,
  y,
  title_txt = "",
  sub_title = "",
  caption_txt = "",
  multi_var_for_color = T,
  color_var = "source",
  show_error_bar = T,
  error_bar_height = 0.2
)
}
\arguments{
\item{df}{data frame with wind speed information}

\item{x}{X-axis variable name}

\item{y}{Y-axis variable name}

\item{title_txt}{plot title}

\item{sub_title}{plot subtitle}

\item{caption_txt}{plot caption name}

\item{multi_var_for_color}{whether or not to use multiple colors (with each variable/column using different colors)}

\item{color_var}{the variable (i.e., column) name to be used for different colors.}

\item{show_error_bar}{show error bar or not}

\item{error_bar_height}{the height of both the ends of the error bar.}
}
\value{

}
\description{
Plot height (across different vertical levels) versus wind speed with error bars.
}
\note{
the error bar is horizontally drawn, i.e., a flip version of a typical error bar plot.
}
