% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{uv_to_meteor_wdir}
\alias{uv_to_meteor_wdir}
\title{Calculate meteorological wind direction}
\usage{
uv_to_meteor_wdir(u, v)
}
\arguments{
\item{u}{U component of the wind vector}

\item{v}{V component of the wind vector}
}
\value{
meteorological wdir
}
\description{
Calculate meteorological wind direction
}
\note{
This function should result in the same wind direction as uv_to_wspd_wdir.
}
