% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{uv_to_wspd_wdir}
\alias{uv_to_wspd_wdir}
\title{Calculate wind speed and direction from U and V winds}
\usage{
uv_to_wspd_wdir(u, v)
}
\arguments{
\item{u}{U component of the wind vector}

\item{v}{V component of the wind vector}
}
\value{

}
\description{
Calculate wind speed and direction from U and V winds
}
